Add tinyPS logo for BIMI
Per the instructions https://bimigroup.org/implementation-guide/
Once this file is hosted, I will update the DNS repo with a BIMI record that will point to where this is hosted. It will not be picked up by providers until our DMARC record moves to quarantine in 2 weeks (also a DNS repo change).