Fix BIMI SVG (#6)
## What was changed
I checked the output of the [new BIMI
tester](https://bimigroup.org/bimi-generator/) and it told me I had to
get rid of x, y, overflow, rect, and change the value of baseprofile. I
did all of those things.

## Why?
Our BIMI logo wasn't showing up and I previously had no information as
to why. This should fix it and make it finally appear.

## Checklist

How was this tested:
Testing not required because it already was not showing up. Making these
changes and looking for it to work is what will be the test.